TruthGuard- Fake News Detection System/requirement.txt
# Core Packages
python>=3.8
streamlit>=1.22.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
nltk>=3.8.0

# Database
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.0

# URL Processing
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
requests>=2.28.0
urllib3>=1.26.0

# NLP/ML
joblib>=1.2.0
spacy>=3.5.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl

# Deployment
gunicorn>=20.1.0
python-dotenv>=0.21.0

# Development
pytest>=7.2.0
black>=23.1.0
flake8>=6.0.0